+++ include/neuralnetwork/neuralnetwork.h
@@ -60,7 +60,7 @@ public:
 
   // Multiple Output layers
   std::vector<NeuralNetworkHelperMetrics> calculate_forecast_metric_all_layers(ErrorCalculation::type error_type) const;
-  std::vector<std::vector<NeuralNetworkHelperMetrics>> calculate_forecast_metrics_all_layers(const std::vector<ErrorCalculation::type>& error_types, bool in_sample = true) const;
+  std::vector<std::vector<NeuralNetworkHelperMetrics>> calculate_forecast_metrics_all_layers(const std::vector<ErrorCalculation::type>& error_types, bool in_sample = true, bool force_checking_indexes = false) const;
 
   double get_learning_rate() const noexcept;
   double get_temperature() const noexcept;
@@ -113,7 +113,7 @@ private:
   double calculate_learning_rate_warmup(int epoch, double completed_percent) const;
 
   std::vector<NeuralNetworkHelperMetrics> calculate_forecast_metrics_impl(const std::vector<ErrorCalculation::type>& error_types, bool in_sample, const Layers* layers) const;
-  std::vector<std::vector<NeuralNetworkHelperMetrics>> calculate_forecast_metrics_all_layers_impl(const std::vector<ErrorCalculation::type>& error_types, bool in_sample, const Layers* layers) const;
+  std::vector<std::vector<NeuralNetworkHelperMetrics>> calculate_forecast_metrics_all_layers_impl(const std::vector<ErrorCalculation::type>& error_types, bool in_sample, const Layers* layers, bool force_checking_indexes = false) const;
 
   std::shared_ptr<NeuralNetworkHelper> create_initial_neural_network_helper(int number_of_epoch, const std::vector<std::vector<double>>& training_inputs, const std::vector<std::vector<double>>& training_outputs) const;